Add files via upload
his is a website developed by me. I used the Bootstrap library to style the elements and add some visual effects. I also wrote a simple JavaScript code to create a button that takes the user back to the homepage.
Along with the website, I included all the used files, including images and a screen recording showing the final version of the website in action.